2022/advent12/advent12_test.go
package main

import "testing"

var testinput string = `Sabqponm
abcryxxl
accszExk
acctuvwj
abdefghi`

func Test_run1(t *testing.T) {
	type args struct {
		inputText string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: "example", args: args{testinput}, want: 31},
		{name: "real", args: args{inputText}, want: 528},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := run1(tt.args.inputText); got != tt.want {
				t.Errorf("run1() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_run2(t *testing.T) {
	type args struct {
		inputText string
	}
	tests := []struct {
		name string
		args args
		want float64
	}{
		{name: "example", args: args{testinput}, want: 29},
		{name: "real", args: args{inputText}, want: 522},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := run2(tt.args.inputText); got != tt.want {
				t.Errorf("run2() = %v, want %v", got, tt.want)
			}
		})
	}
}
